LF-4378 Restore manual control of cancel modal to form variant; removed SingleStep component as no longer needed; remove remounting key to get modal-controlling useEffect to trigger
