+++ packages/webapp/src/containers/Animals/SingleAnimalView/index.tsx
@@ -110,7 +110,6 @@ function SingleAnimalView({ isCompactSideMenu, history }: AddAnimalsProps) {
           cancelModalTitle={t('ANIMALS.EDIT_ANIMAL_FLOW')}
           isEditing={isEditing}
           setIsEditing={setIsEditing}
-          key={isEditing ? 'edit' : 'readonly'}
         />
       )}
       {/* TODO: Has not yet been scoped. Revisit after movement */}
